+++ DashMart/Core/HomeScreen/HomeScreen.swift
@@ -31,9 +31,13 @@ struct HomeScreen: View {
         Int(UIScreen.main.bounds.width) / 67
     }
     
-    @State private var isShowingFilters = false 
+    @State private var isShowingFilters = false
     @State private var isButtonActive = false
     @State private var filtersApplied = false
+    @State private var minPrice: Double?
+    @State private var maxPrice: Double?
+    @State private var sortType: SortType = .none
+    @State private var filterText = ""
     
     var body: some View {
         VStack(spacing: 16) {
@@ -89,40 +93,103 @@ struct HomeScreen: View {
                     }
                 }
                 .padding(.horizontal, 20)
-            
-                TitleFilters(text: "Products", action: {
-                    isShowingFilters.toggle()
-                }, filtersApplied: $filtersApplied, isButtonActive: $isButtonActive)
-                .bottomSheet(isPresented: $isShowingFilters, detents: [.medium()]) {
-                    FilterProductsVM(showAlphabeticalSort: true)
-                }
-                    .padding(.horizontal, 20)
                 
-                ScrollView {
-                    LazyVGrid(
-                        columns: [.init(),.init()],
-                        spacing: 8
-                    ) {
-                        ForEach(filteredProducts) { product in
+                VStack {
+                    TitleFilters(text: "Products", action: {
+                        isShowingFilters.toggle()
+                    }, filtersApplied: $filtersApplied, isButtonActive: $isButtonActive)
+                    .padding(.horizontal, 20)
+                    .bottomSheet(isPresented: $isShowingFilters, detents: [.medium()]) {
+                        VStack(spacing: 16) {
+                            Text("Filter Products")
+                                .font(.system(size: 16))
+                                .foregroundStyle(Color(hex: "#393F42"))
+                                .padding()
+                            Picker(selection: $sortType, label: Text("Sort by")) {
+                                Text("None").tag(SortType.none)
+                                Text("A-Z").tag(SortType.alphabeticalAscending)
+                                Text("Z-A").tag(SortType.alphabeticalDescending)
+                            }
+                            .pickerStyle(SegmentedPickerStyle())
+                            .padding()
+                            
+                            HStack {
+                                Text("Price")
+                                    .font(.system(size: 16))
+                                    .foregroundStyle(Color(hex: "#393F42"))
+                                Spacer()
+                            }
+                            .padding(.horizontal)
+                            HStack {
+                                TextField("Min", text: Binding<String>(
+                                    get: { minPrice.map { String($0) } ?? "" },
+                                    set: { minPrice = Double($0) }
+                                ))
+                                .textFieldStyle(RoundedBorderTextFieldStyle())
+                                
+                                Image(systemName: "ellipsis")
+                                
+                                Spacer()
+                                TextField("Max", text: Binding<String>(
+                                    get: { maxPrice.map { String($0) } ?? "" },
+                                    set: { maxPrice = Double($0) }
+                                ))
+                                .textFieldStyle(RoundedBorderTextFieldStyle())
+                            }
+                            .padding(.horizontal)
+                            .padding(.bottom, 40)
+                        }
+                        HStack {
                             Button(
                                 action: {
-                                    selectedProduct = product
-                                    isShowingDetails = true
-                                },
-                                label: {
-                                    ProductItem(
-                                        product: product,
-                                        storage: storage,
-                                        showWishlistButton: false
-                                    )
+                                    clearFilters()
+                                }, label: {
+                                    Text("Clear Filter")
+                                        .foregroundStyle(Color(hex: "#E53935"))
+                                        .modifier(DashRoundedTitle(style: .gray))
+                                }
+                            )
+                            
+                            Button(
+                                action: {
+                                    applyFilters()
+                                }, label: {
+                                    Text("Apply")
+                                        .modifier(DashRoundedTitle())
                                 }
                             )
                         }
+                        .padding(.horizontal)
+                        .background(Color.white)
                     }
-                    .padding(.horizontal, 20)
                 }
             }
+            
+            ScrollView {
+                LazyVGrid(
+                    columns: [.init(),.init()],
+                    spacing: 8
+                ) {
+                    ForEach(filteredProducts) { product in
+                        Button(
+                            action: {
+                                selectedProduct = product
+                                isShowingDetails = true
+                            },
+                            label: {
+                                ProductItem(
+                                    product: product,
+                                    storage: storage,
+                                    showWishlistButton: false
+                                )
+                            }
+                        )
+                    }
+                }
+                .padding(.horizontal, 20)
+            }
         }
+        
         .padding(.top)
         .task {
             loading = true
@@ -154,6 +221,66 @@ struct HomeScreen: View {
         .bottomSheet(isPresented: $isShowingLocation, detents: [.medium()]) {
             CountrySelection()
         }
+    }
+    enum SortType {
+        case none
+        case alphabeticalAscending
+        case alphabeticalDescending
+    }
+    
+    func applyFilters(closeBottomSheet: Bool = true) {
+        filteredProducts = products
+        
+        var isFilterApplied = false
+        
+        if !filterText.isEmpty {
+            filteredProducts = filteredProducts.filter { $0.title.localizedCaseInsensitiveContains(filterText) }
+            isFilterApplied = true
+        }
+        
+        if let minPrice = minPrice {
+            filteredProducts = filteredProducts.filter { $0.price >= minPrice }
+            isFilterApplied = true
+        }
+        
+        if let maxPrice = maxPrice {
+            if let minPrice = minPrice, maxPrice < minPrice {
+                self.maxPrice = minPrice
+            }
+            filteredProducts = filteredProducts.filter { $0.price <= maxPrice }
+            isFilterApplied = true
+        }
+        
+        guard isFilterApplied || sortType != .none else {
+            return
+        }
+        switch sortType {
+        case .alphabeticalAscending:
+            filteredProducts.sort { $0.title < $1.title }
+        case .alphabeticalDescending:
+            filteredProducts.sort { $0.title > $1.title }
+        default:
+            break
+        }
+        
+        isButtonActive = !filteredProducts.isEmpty
+        
+        if closeBottomSheet {
+            isShowingFilters = false
+        }
+        
+        filtersApplied = true
+        
+    }
+    
+    func clearFilters() {
+        filterText = ""
+        minPrice = nil
+        maxPrice = nil
+        sortType = .none
+        applyFilters(closeBottomSheet: false)
+        isButtonActive = false
+        filtersApplied = false
     }
 }
